32 bit data type for scan interval and scan window
The fields in the struct nrf_radio_request_earliest_t in nrf_soc.h are 32 bit long so even the values in btle.h have to be 32 bit long otherwise the behaviour is not correct. I have checked it looking at the power profile while scanning and noticed that the interval between scans was not correct.
I haven't checked for the advertising but I expect a similar problem.